requirements.txt
Flask>=2.3.2
GitPython>=3.1.30